# Don't use the LLVM symbolizer to look up symbols.