@ltratt 2022-08-15
Out of curiosity, if you do profile an instance (with `perf` or `valgrind` or whatever) how much time is spent in `vaddr_to_sym_and_obj` (and its descendents)? If it is substantial, maybe we should just do the interval tree thing now?